Refactor Substance Painter utilities
- Extracted `texture_set_name` and `docs_link_html` to `pipe.sp.util`.
- Added `pipe.sp.houdini` for Houdini asset builder integration.
- Replaced redundant implementations with shared utilities.
- Updated imports and references across modules.
