+++ pipeline/pipe/sp/__init__.py
@@ -1,12 +1,24 @@
 # import substance painter-specific scripts
-from . import assetfile, channels, dialogs, export, local, metadata, reload
+from . import (
+    assetfile,
+    channels,
+    dialogs,
+    export,
+    houdini,
+    local,
+    metadata,
+    reload,
+    util,
+)
 
 __all__ = [
     "assetfile",
     "channels",
     "dialogs",
     "export",
+    "houdini",
     "local",
     "metadata",
     "reload",
+    "util",
 ]
